requirements.txt
streamlit==1.29.0
python-dotenv==1.0.0
PyPDF2==3.0.1
langchain==0.0.350
openai==0.28.1
faiss-cpu==1.8.0
langchain-community==0.0.10
tiktoken